+++ easygui_qt/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 __author__ = 'André Roberge'
 __email__ = 'andre.roberge@gmail.com'
-__version__ = '0.3.0'
+__version__ = '0.4.0'
 
 from .easygui_qt import *
 from .easygui_qt import __all__